% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Prompt for User Input}
\usage{
ask(prompt, default = NULL)
}
\arguments{
\item{prompt}{prompt to the user}

\item{default}{default value (shown to the user)}
}
\value{
character value of the user's response
}
\description{
Prompt the user to enter some text with or without a default.
In the context of this package it is used to prompt for settings
needed to access the sftp site.
}
\author{
Howard J. Seltman \email{hseltman@stat.cmu.edu} and Francis R. Kovacs
}
