% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull Files from an Sftp Server}
\usage{
pull(files)
}
\arguments{
\item{files}{a character vector containing the files to be loaded}
}
\value{
None
}
\description{
Pull one or more files from the main sftp folder into the
working directory.
}
\details{
Students use pull(files) to download file(s) to their computer from
the sftp site.  One or more files can specified.  Files are loaded to the
student's working directory.  If the file already exists the user is
asked if s(he) wants to overwrite the old file.  If "n"o is chosen,
the user is asked for a new file name.
}
\references{
https://stackoverflow.com/questions/15950396/sftp-get-files-with-r
https://jonkimanalyze.wordpress.com/2014/11/20/r-quick-sftp-file-transfer/
}
\author{
Howard J. Seltman \email{hseltman@stat.cmu.edu} and Francis R. Kovacs
}
