% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Push Files to an Sftp Server}
\usage{
push(files)
}
\arguments{
\item{files}{a character vector containing the files to be uploaded.}
}
\value{
None
}
\description{
Push one or more files from to the student's working directory (or any
other student director) to the student's directory on the sftp server.
}
\references{
https://stackoverflow.com/questions/15950396/sftp-get-files-with-r
https://jonkimanalyze.wordpress.com/2014/11/20/r-quick-sftp-file-transfer/
}
\author{
Howard J. Seltman \email{hseltman@stat.cmu.edu} and Francis R. Kovacs
}
