% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\alias{pull}
\title{Reset}
\usage{
: 
Students are attempting to code something, and you want everyone to
   "catch up" to your version.  Run push("myfile.R").  Each student can
   then run pull("myFile.R"), and open your version on their computer.
}
\description{
This function allows the user to reset the configuration for their ftp server.
}
\details{
:
Students use push(files) to upload file(s) from their computer to
  the sftp site in their folder (as specified by their setup user name).
  Typically one or more files in the current folder are specified, but
  relative and absolute locations are allowed.
  Students use pull(files) to download files to their current folder
  from the root folder of the ftp site.
  sftp site's root folder (by specifying user name ".").
  Teachers use pull(files, who="studentUserId") to download one or more
  student files to their current folder.
}
\references{
https://stackoverflow.com/questions/15950396/sftp-get-files-with-r
https://jonkimanalyze.wordpress.com/2014/11/20/r-quick-sftp-file-transfer/
}
\author{
Francis R. Kovacs
}
