% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup the User to Access the Sftp Site}
\usage{
setup()
}
\value{
None
}
\description{
Each user will have their own subfolder on the sftp site to
  which they upload.  This function sets the user's info 
  in a persistent data object called 'userSftpInfo'.
  It will be called during .onLoad and whenever the user
  re-runs it.
}
\author{
Howard J. Seltman \email{hseltman@stat.cmu.edu} and Francis R. Kovacs
}
