% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftpSetup.R
\name{sftpSetup}
\alias{sftpSetup}
\title{Setup the User to Access the Sftp Site}
\usage{
sftpSetup()
}
\value{
None
}
\description{
Each user will have their own subfolder on the sftp site to
  which they upload.  This function sets the site and user info 
  in a persistent file called "pushPullInfo.txt" in the packages
  root directory.  The information comprises the sftp URL, the sftp
  user name, the sftp password, and the user's unique user name.
  (The intent is to allow a simple setup where the sftp user name and
  password are shared across students.)  When the library is loaded,
  the information is loaded from that file and place in an option()
  called 'pushPullInfo'.
}
\details{
The file "pushPullInfo.txt" is a plain text file with four lines.  The
  lines contain the sftp site URL (without the "sftp://" prefix, and possibly
  with a "/directory" suffix), the sftp user name, the sftp password,
  and the users unique id (e.g., base portion of University email address).

The return value of getOption("pushPullInfo") is a named character vector
  with elements "sftpSite", "sftpName", "sftpPassword", and "userName".
}
\author{
Howard J. Seltman \email{hseltman@stat.cmu.edu} and Francis R. Kovacs
}
