% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Ask}
\usage{
: 
After called for each variable, the resulting dataframe will be saved to sysdata.rda
}
\description{
A function that prompts for a given variable name. In the context 
    of this package it is used to prompt for settings.
}
\author{
Howard J. Seltman
}
