% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull Files from an Sftp Server}
\usage{
pull(files, who = NULL)
}
\arguments{
\item{files}{a character vector containing the files to be loaded}

\item{who}{a length-1 character vector specifying the folder on
the sftp server from which the file is read}
}
\value{
None
}
\description{
Pull one or more files from the main sftp folder into the
working directory.  If \code{"who"} is used, the file is
read from that subfolder.
}
\details{
Students use push(files) to upload file(s) from their computer to
the sftp site in their folder (as specified by their setup user name).
Typically one or more files in the current folder are specified, but
relative and absolute locations are allowed.
Students use pull(files) to download files to their current folder
from the root folder of the ftp site.

Teachers use push(files) to upload file(s) from their computer to the
sftp site's root folder (by specifying user name ".").
Teachers use pull(files, who="studentUserId") to download one or more
student files to their current folder.
}
\references{
https://stackoverflow.com/questions/15950396/sftp-get-files-with-r
https://jonkimanalyze.wordpress.com/2014/11/20/r-quick-sftp-file-transfer/
}
\author{
Howard J. Seltman \email{hseltman@stat.cmu.edu} and Francis R. Kovacs
}
