% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Push}
\usage{
: 
1. Students are working in class and a student wants to know why their code
   failed or you want to show a working (or not working) example from a
   particular student.  The student pushes the file with push("myFile.R")
   and you pull it with pull("myFile.R", who="studentId").  Now you can
   open the file in your current R session.
}
\description{
This is a teaching tool to encourage interaction in programming classes. 
    It is based on the existance of an FTP server where files are shared and exchanged.
}
\details{
:
Students use push(files) to upload file(s) from their computer to
  the sftp site in their folder (as specified by their setup user name).
  Typically one or more files in the current folder are specified, but
  relative and absolute locations are allowed.
  Students use pull(files) to download files to their current folder
  from the root folder of the ftp site.
  sftp site's root folder (by specifying user name ".").
  Teachers use pull(files, who="studentUserId") to download one or more
  student files to their current folder.
}
\references{
https://stackoverflow.com/questions/15950396/sftp-get-files-with-r
https://jonkimanalyze.wordpress.com/2014/11/20/r-quick-sftp-file-transfer/
}
\author{
Howard J. Seltman and Francis R. Kovacs
}
