% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Configure the sftp server information}
\usage{
: This function will be called during .onLoad and whenever the user prompts it.
}
\description{
This function allows the user to reset the configuration for their ftp server.
}
\details{
: To do later.
}
\author{
Francis R. Kovacs
}
